refactor: 1단계 코드 패키지 구조 정리
