+++ app/src/main/java/nextstep/payments/new_card/NewCardTopBar.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.new_card
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
@@ -10,6 +10,8 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import nextstep.payments.ui.theme.PaymentsTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -38,4 +40,12 @@ fun NewCardTopBar(
         },
         modifier = modifier
     )
+}
+
+@Preview
+@Composable
+private fun NewCardTopBarPreview() {
+    PaymentsTheme {
+        NewCardTopBar(onBackClick = {}, onSaveClick = {})
+    }
 }
